+++ Classes/Domain/ValueObject/ChatMessage.php
@@ -9,9 +9,9 @@ declare(strict_types=1);
 
 namespace Netresearch\NrLlm\Domain\ValueObject;
 
-use InvalidArgumentException;
 use JsonSerializable;
 use Netresearch\NrLlm\Domain\Enum\MessageRole;
+use Netresearch\NrLlm\Exception\InvalidArgumentException;
 
 /**
  * Value Object representing a chat message.