+++ Classes/Exception/ConfigurationNotFoundException.php
@@ -14,4 +14,4 @@ use RuntimeException;
 /**
  * Exception thrown when an LLM configuration cannot be found.
  */
-class ConfigurationNotFoundException extends RuntimeException {}
+class ConfigurationNotFoundException extends RuntimeException implements NrLlmExceptionInterface {}