+++ Classes/Exception/PromptTemplateNotFoundException.php
@@ -14,4 +14,4 @@ use RuntimeException;
 /**
  * Exception thrown when a requested prompt template is not found.
  */
-class PromptTemplateNotFoundException extends RuntimeException {}
+class PromptTemplateNotFoundException extends RuntimeException implements NrLlmExceptionInterface {}